+++ 유니_이혜윤/3주차 추가미션/src/Header.jsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+const Container = styled.div`
+  width: 100%;
+  height: 20px;
+  display: flex;
+  position: fixed;
+  top: 0; left: 0;
+  background-color: #181935;
+  z-index: 1000;
+`;
+
+const Header = () => {
+  return (
+    <Container>
+
+    </Container>
+  )
+}
+
+export default Header;
