+++ 유니_이혜윤/3주차 추가미션/src/index.css
@@ -1,4 +1,3 @@
 * {
-  max-width: 1280px;
   text-align: center;
 }
